requirements.txt
# Core dependencies
# CRITICAL: torch>=2.4.0 required for native bfloat16 support (Wan2.2)
torch>=2.4.0
torchvision>=0.19.0
diffusers>=0.30.0  # Required for WanPipeline and AutoencoderKLWan
transformers>=4.40.0
accelerate>=0.30.0

# ONNX and Optimization
onnx>=1.15.0
onnxruntime-gpu>=1.16.0
optimum[exporters]>=1.16.0

# TensorRT (install separately via NVIDIA)
# tensorrt>=8.6.0
# pycuda>=2022.2

# Utilities
numpy>=1.24.0
pillow>=10.0.0
tqdm>=4.65.0
pyyaml>=6.0
opencv-python>=4.8.0

# Benchmarking and Validation
scikit-image>=0.21.0
lpips>=0.1.4
pytorch-fid>=0.3.0

# Optional quantization
# bitsandbytes>=0.41.0

# Development
pytest>=7.4.0
black>=23.0.0

